pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "portaler"
version = "0.0.1"
authors = [
    { name = "JiffyRob" },
    { name = "YeatsAiden" },
    { name = "gresm" },
    { name = "AndreyVarvar" },
    { name = "Matiiss" },
    { name = "aatle" },
]
description = ""
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = ["pygame-ce"]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.15.0",
]

[project.urls]
Homepage = "https://github.com/pygame-examples/legacy-code-01-team-merge-mishaps"
Issues = "https://github.com/pygame-examples/legacy-code-01-team-merge-mishaps/issues"

[tool.pylint.main]
disable = ["all"]
enable = ["attribute-defined-outside-init"]

[tool.ruff]
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "I", "C901"]

[tool.ruff.lint.isort]
combine-as-imports = true
